refactor more by reducing cross import
