+++ rust/sglang-server/src/message.rs
@@ -15,10 +15,12 @@ pub use egress::{
     frame_egress_result,
 };
 pub use request::{
-    ControlRequest, GenerateBody, GenerateRequest, IngressMsg, RequestKind, abort_req_msgpack,
+    ControlRequest, GenerateBody, GenerateRequest, RequestKind, abort_req_msgpack,
     control_req_msgpack,
 };
 pub use sampling::normalize_sampling_params;
+
+use bytes::Bytes;
 
 use crate::fsm::RequestState;
 use crate::ids::RidHash;
@@ -37,4 +39,42 @@ pub struct Request {
     pub sink: EgressSink,
     /// Discriminant + variant body (generate vs control).
     pub kind: RequestKind,
+}
+
+/// One ingress-ring entry, split columnar: the scalar `header` (msgpack, `input_ids`
+/// omitted) + the raw int64 `ids` cell, so the big tensor never goes through msgpack.
+#[derive(Debug)]
+pub struct IngressMsg {
+    pub header: Bytes,
+    pub ids: Bytes,
+}
+
+/// Messages to a Detokenizer shard. `Register` carries the per-request sink for
+/// the shard's local `rid_hash -> sink` map; everything routes by `RidHash::shard`.
+pub enum DetokMsg {
+    Register {
+        rid_hash: RidHash,
+        /// Client-visible rid string — kept in `DetokState` so the shard can
+        /// emit `TmEvent::Abort(rid)` (the wire needs the string, not the hash).
+        rid: String,
+        sink: EgressSink,
+        /// Decode logprob token ids to text here (CPU-bound) not on the api threads.
+        decode_logprob_text: bool,
+        /// `SamplingParams.no_stop_trim`: keep the matched stop; default trims it.
+        no_stop_trim: bool,
+    },
+    /// One decode step's chunks for *this shard*. Batched because `tm-egress` blocks
+    /// per send, so one message per request cost ~1.3 µs × batch (5.1x at 4096).
+    Chunks(Vec<ChunkEvent>),
+    /// Control result: one already-serialized payload delivered to the sink verbatim.
+    Result {
+        rid_hash: RidHash,
+        payload: bytes::Bytes,
+    },
+    /// Terminal per-request failure → an `Error` to the sink (a 400, not a crash).
+    Fail { rid_hash: RidHash, message: String },
+    /// Drop the `rid_hash -> sink` entry for a request rejected before the scheduler
+    /// (the rejecting stage already answered the client); else `Register` leaks one
+    /// entry.
+    Deregister { rid_hash: RidHash },
 }
